# Conflicts:
#	app/src/main/java/com/example/reviewmate/AddActivity.kt
#	app/src/main/java/com/example/reviewmate/AuthActivity.kt
#	app/src/main/java/com/example/reviewmate/FragmentFive.kt
#	app/src/main/java/com/example/reviewmate/FragmentFive_ReviewList.kt
#	app/src/main/java/com/example/reviewmate/FragmentFour.kt
#	app/src/main/java/com/example/reviewmate/FragmentOne.kt
#	app/src/main/java/com/example/reviewmate/FragmentThree.kt
#	app/src/main/java/com/example/reviewmate/FragmentTwo.kt
#	app/src/main/java/com/example/reviewmate/ItemFeedModel.kt
#	app/src/main/java/com/example/reviewmate/MainActivity.kt
#	app/src/main/java/com/example/reviewmate/MovieAdapter.kt
#	app/src/main/java/com/example/reviewmate/MovieDetailsActivity.kt
#	app/src/main/java/com/example/reviewmate/MyApplication.kt
#	app/src/main/java/com/example/reviewmate/MyFeedAdapter.kt
#	app/src/main/java/com/example/reviewmate/ReviewDetailActivity.kt
#	app/src/main/java/com/example/reviewmate/UserModel.kt
#	app/src/main/java/com/example/reviewmate/common/Movie.kt
#	app/src/main/java/com/example/reviewmate/common/MoviesRepository.kt
#	app/src/main/java/com/example/reviewmate/common/TMDBClient.kt
#	app/src/main/res/layout/activity_add.xml
#	app/src/main/res/layout/activity_review_detail.xml
#	app/src/main/res/layout/fragment_five.xml
#	app/src/main/res/layout/fragment_five__review_list.xml
#	app/src/main/res/layout/fragment_four.xml
#	app/src/main/res/layout/fragment_one.xml
#	app/src/main/res/layout/item_feed.xml
#	app/src/main/res/layout/item_movie.xml
#	app/src/main/res/layout/movie_detail.xml
#	app/src/main/res/menu/menu_list.xml

+++ app/src/main/java/com/example/reviewmate/MyFeedAdapter.kt
@@ -1,36 +1,42 @@
 package com.example.reviewmate
 
+import android.annotation.SuppressLint
 import android.app.AlertDialog
-import android.content.*
-import android.graphics.BitmapFactory
+import android.content.ClipData
+import android.content.Context
+import android.content.DialogInterface
+import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
 import android.widget.Toast
+import androidx.core.content.ContentProviderCompat
 import androidx.core.content.ContentProviderCompat.requireContext
+import androidx.core.content.ContextCompat.startActivity
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.reviewmate.MyApplication
+import com.example.reviewmate.MyApplication.Companion.auth
+import com.example.reviewmate.MyApplication.Companion.checkAuth
 import com.example.reviewmate.MyApplication.Companion.db
-import com.example.reviewmate.MyApplication.Companion.getImageUrl
 import com.example.reviewmate.MyApplication.Companion.storage
 import com.example.reviewmate.databinding.ItemFeedBinding
 import com.google.android.gms.tasks.Task
-import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.ktx.auth
+import com.google.firebase.firestore.DocumentReference
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.ktx.Firebase
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
+import kotlinx.coroutines.tasks.await
 import okhttp3.internal.notifyAll
 
 class MyFeedViewHolder(val binding: ItemFeedBinding) : RecyclerView.ViewHolder(binding.root)
 
-class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedModel>, val sourceFragment: String): RecyclerView.Adapter<MyFeedViewHolder>() {
+class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedModel>): RecyclerView.Adapter<MyFeedViewHolder>() {
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyFeedViewHolder {
         val layoutInflater = LayoutInflater.from(parent.context)
@@ -41,7 +47,13 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
         return itemList.size
     }
 
+    private fun removeItem(position: Int) {
+        itemList.removeAt(position)
+        notifyItemRemoved(position)
+    }
+
     override fun onBindViewHolder(holder: MyFeedViewHolder, position: Int) {
+        val position = holder.getAdapterPosition()
         val data = itemList.get(position)
 
         holder.binding.run {
@@ -51,7 +63,7 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
             itemDateView.text = data.date
             itemMovieView.text = data.movie
             itemRateView.text = data.rate
-            // ===========================
+
             // Firestore에서 user 모델 데이터 가져오기
             val db = FirebaseFirestore.getInstance()
             val usersCollection = db.collection("users")
@@ -72,7 +84,33 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
                     Toast.makeText(context, "${profileImageUrl}", Toast.LENGTH_SHORT).show()
                 }
 
-                // =============================
+
+//            usersCollection.whereEqualTo("userEmail", userEmail)
+//                .get()
+//                .addOnSuccessListener { querySnapshot ->
+//                    if (!querySnapshot.isEmpty) {
+//
+//                        val userDocument = querySnapshot.documents[0]
+//                        profileImageUrl = userDocument.getString("imageUrl")
+//
+//                        if( profileImageUrl != null){
+//                            // Glide를 사용하여 프로필 이미지 로드
+//                            Glide.with(context)
+//                                .load(profileImageUrl)
+//                                .into(holder.binding.itemImg)
+//                        }
+//                    }
+//                }
+//                .addOnFailureListener { exception ->
+//                    Log.e("MyFeedAdapter", "Error getting user document: $exception")
+//                }
+
+                if (itemContentView.text.isNotEmpty()) {
+                    itemContentView.visibility = View.VISIBLE
+                } else {
+                    itemContentView.visibility = View.GONE
+                }
+
                 itemTitleView.setOnClickListener {
                     val bundle: Bundle = Bundle()
                     bundle.putString("email", data.email)
@@ -83,8 +121,9 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
                     bundle.putString("rate", data.rate)
                     bundle.putString("userEmail", data.email)
                     bundle.putString("date", data.date)
-                    bundle.putString("image_url", profileImageUrl)
                     bundle.putString("movieImage", data.movieImage)
+                    bundle.putString("reviewId", data.docId)
+                    bundle.putString("image_url", profileImageUrl)
 
                     Intent(context, ReviewDetailActivity::class.java).apply {
                         putExtras(bundle)
@@ -96,26 +135,47 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
                     override fun onClick(dialog: DialogInterface?, which: Int) {
                         when (which) {
                             DialogInterface.BUTTON_POSITIVE -> {
-                                storage.getReference().child("images").child("${data.docId!!}.jpg")
-                                    .delete()
+                                if (data.docId !== null) {
+                                    storage.getReference().child("images")
+                                        .child("${data.docId!!}.jpg")
+                                        .delete()
+                                    db.collection("reviews").document("${data.docId}")
+                                        .delete()
+                                        .addOnSuccessListener {
+                                            Toast.makeText(
+                                                context,
+                                                "삭제가 완료되었습니다.",
+                                                Toast.LENGTH_SHORT
+                                            ).show()
 
-                                db.collection("reviews").document("${data.docId}")
-                                    .delete()
-                                    .addOnSuccessListener {
-                                        Toast.makeText(
-                                            context,
-                                            "삭제가 완료되었습니다.",
-                                            Toast.LENGTH_SHORT
-                                        ).show()
-                                    }
-                                    .addOnFailureListener {
-                                        Toast.makeText(
-                                            context,
-                                            "삭제가 실패하였습니다.",
-                                            Toast.LENGTH_SHORT
-                                        ).show()
-                                    }
-                                notifyItemRemoved(position)
+                                            val userDocRef = MyApplication.db.collection("users")
+                                                .document(auth.uid.toString())
+                                            MyApplication.db.collection("users")
+                                                .document("${MyApplication.auth.uid}")
+                                                .get()
+                                                .addOnSuccessListener { documentSnapshot ->
+                                                    if (documentSnapshot.exists()) {
+                                                        val currentCount =
+                                                            documentSnapshot.getLong("userReviewCount")
+                                                        currentCount?.let {
+                                                            val updatedCount = it - 1
+                                                            updateCount(userDocRef, updatedCount)
+                                                            removeItem(holder.adapterPosition)
+                                                        }
+                                                    }
+                                                }
+                                        }
+                                        .addOnFailureListener {
+                                            Toast.makeText(
+                                                context,
+                                                "삭제가 실패하였습니다.",
+                                                Toast.LENGTH_SHORT
+                                            ).show()
+                                        }
+                                } else {
+                                    Toast.makeText(context, "문서가 존재하지 않습니다.", Toast.LENGTH_SHORT)
+                                        .show()
+                                }
                             }
                             DialogInterface.BUTTON_NEGATIVE -> {
                                 Log.d("ToyProject", "DialogInterface.BUTTON_NEGATIVE")
@@ -124,31 +184,23 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
                     }
                 }
 
-
-                if (sourceFragment.equals("FragmentOne")) {
-                    reviewDelete.visibility = View.INVISIBLE
-                    itemImg.visibility = View.GONE
-                    itemEmailView.visibility = View.GONE
-                    itemDateView.visibility = View.GONE
-                } else if (sourceFragment.equals("FragmentTwo")) {
-                    if (itemEmailView.text == MyApplication.email) {
-                        reviewDelete.visibility = View.VISIBLE
+                if (data.email == MyApplication.email) {
+                    reviewDelete.visibility = View.VISIBLE
+                    reviewDelete.setOnClickListener {
+                        AlertDialog.Builder(context).run {
+                            setTitle("정말 삭제하시겠습니까?")
+                            setMessage("한 번 삭제하면 되돌릴 수 없습니다.")
+                            setNegativeButton("Cancle", alertHandler)
+                            setPositiveButton("Yes", alertHandler)
+                            show()
+                        }
                     }
-                }
-
-
-                reviewDelete.setOnClickListener {
-                    AlertDialog.Builder(context).run {
-                        setTitle("정말 삭제하시겠습니까?")
-                        setMessage("한 번 삭제하면 되돌릴 수 없습니다.")
-                        setNegativeButton("Cancle", alertHandler)
-                        setPositiveButton("Yes", alertHandler)
-                        show()
-                    }
+                } else {
+                    reviewDelete.visibility = View.GONE
                 }
             }
 
-            //스토리지 이미지 다운로드……………………
+            //스토리지 이미지 다운로드........................
             val imageRef = MyApplication.storage.reference.child("images/${data.docId}.jpg")
             imageRef.downloadUrl.addOnCompleteListener { task ->
                 if (task.isSuccessful) {
@@ -159,7 +211,19 @@ class MyFeedAdapter(val context: Context, val itemList: MutableList<ItemFeedMode
                 }
             }
         }
+    }
+    fun updateCount(docRef: DocumentReference, updatedValue: Long) {
+        val updates = hashMapOf<String, Any>(
+            "userReviewCount" to updatedValue
+        )
+
+        docRef.update(updates)
+            .addOnSuccessListener {
+                // 업데이트 성공 처리
+            }
+            .addOnFailureListener { e ->
+                // 업데이트 실패 처리
+            }
 
     }
 }
-
